restarting design of code from scratch
