+++ src/main/java/RunTimeTerror/Game.java
@@ -1,0 +1,28 @@
+package RunTimeTerror;
+
+import java.awt.*;
+
+/**
+ * Hello world!
+ *
+ */
+public class Game extends Canvas implements Runnable
+{
+    public static final long serialVersionUID = 4602267202262990821L;
+
+
+    public void run(){
+
+    }
+
+    public synchronized void start(){
+
+    }
+
+    public static void main( String[] args )
+    {
+
+
+    }
+
+}
